pyproject.toml
[project]
name = "pydantic-open-ai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.60.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pytest>=9.0.2",
]
